+++ packages/language-service/src/utils.ts
@@ -6,8 +6,10 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {CompileDirectiveSummary, CompileTypeMetadata, CssSelector, ParseSourceSpan, identifierName} from '@angular/compiler';
+import {AstPath, CompileDirectiveSummary, CompileTypeMetadata, CssSelector, DirectiveAst, ElementAst, EmbeddedTemplateAst, HtmlAstPath, Node as HtmlNode, ParseSourceSpan, RecursiveTemplateAstVisitor, RecursiveVisitor, TemplateAst, TemplateAstPath, identifierName, templateVisitAll, visitAll} from '@angular/compiler';
+import {DiagnosticTemplateInfo} from '@angular/compiler-cli';
 import * as ts from 'typescript';
+
 import {SelectorInfo, TemplateInfo} from './common';
 import {Span} from './types';
 
@@ -109,4 +111,69 @@ export function isTypescriptVersion(low: string, high?: string) {
   if (high && (version.substring(0, high.length) > high)) return false;
 
   return true;
+}
+
+export function diagnosticInfoFromTemplateInfo(info: TemplateInfo): DiagnosticTemplateInfo {
+  return {
+    fileName: info.fileName,
+    offset: info.template.span.start,
+    query: info.template.query,
+    members: info.template.members,
+    htmlAst: info.htmlAst,
+    templateAst: info.templateAst
+  };
+}
+
+export function findTemplateAstAt(
+    ast: TemplateAst[], position: number, allowWidening: boolean = false): TemplateAstPath {
+  const path: TemplateAst[] = [];
+  const visitor = new class extends RecursiveTemplateAstVisitor {
+    visit(ast: TemplateAst, context: any): any {
+      let span = spanOf(ast);
+      if (inSpan(position, span)) {
+        const len = path.length;
+        if (!len || allowWidening || isNarrower(span, spanOf(path[len - 1]))) {
+          path.push(ast);
+        }
+      } else {
+        // Returning a value here will result in the children being skipped.
+        return true;
+      }
+    }
+
+    visitEmbeddedTemplate(ast: EmbeddedTemplateAst, context: any): any {
+      return this.visitChildren(context, visit => {
+        // Ignore reference, variable and providers
+        visit(ast.attrs);
+        visit(ast.directives);
+        visit(ast.children);
+      });
+    }
+
+    visitElement(ast: ElementAst, context: any): any {
+      return this.visitChildren(context, visit => {
+        // Ingnore providers
+        visit(ast.attrs);
+        visit(ast.inputs);
+        visit(ast.outputs);
+        visit(ast.references);
+        visit(ast.directives);
+        visit(ast.children);
+      });
+    }
+
+    visitDirective(ast: DirectiveAst, context: any): any {
+      // Ignore the host properties of a directive
+      const result = this.visitChildren(context, visit => { visit(ast.inputs); });
+      // We never care about the diretive itself, just its inputs.
+      if (path[path.length - 1] == ast) {
+        path.pop();
+      }
+      return result;
+    }
+  };
+
+  templateVisitAll(visitor, ast);
+
+  return new AstPath<TemplateAst>(path, position);
 }
